pyproject.toml
[project]
name = "template"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "asyncmy (>=0.2.10,<0.3.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "openai (>=1.69.0,<2.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "fastapi-pagination (>=0.12.34,<0.13.0)",
    "pydantic-extra-types (>=2.10.3,<3.0.0)",
    "msgspec (>=0.19.0,<0.20.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
pyink = "^24.10.1"
isort = "^6.0.1"

